# Elide bounds checks in uo_hash64_with_seeds