Elide bounds checks in uo_hash64_with_seeds

Write the loop just slightly differently.

With `while s.len() >= 64 {`, the optimizer can elide the bounds
checks for the slicing in the loop, since they follow from `s.len() >= 64`.

Use fully explicit slice ranges, so tha the assertion for length in
fetch64 is optimized out as well.

This improves the benchmark bench_pseurand_big_farm_direct by 15%.