@bluss 2015-08-26
This change turned out to be pretty magical. All bounds checks elided, everything gone in favour of the boundary check that the while loop performs (`s.len() >= 64`).

Profiled with `perf` on linux to hilight hotspots and give disassembly. I noticed all the slicing had boundary checks. The loop does the same thing this way, but the loop condition is much more obvious to the optimizer. “Length >= 64 huh, I can use that”