@bluss 2015-08-26
By the way, big props for writing so much code without unsafe. It seems to perform really well. Profiling and tricks like this are somewhat needed to fix the last pieces.