@bluss 2015-08-26
Benchmarks I see are: (only bench_pseurand_big_farm_direct matters)

before
```
test bench_pseurand_big_farm          ... bench:   2,737,315 ns/iter (+/- 100,759)
test bench_pseurand_big_farm_direct   ... bench:     394,406 ns/iter (+/- 25,857)
test bench_pseurand_big_fnv           ... bench:   3,101,386 ns/iter (+/- 45,235)
test bench_pseurand_big_sip           ... bench:   1,458,863 ns/iter (+/- 24,106)
```

after
```
test bench_pseurand_big_farm          ... bench:   2,586,794 ns/iter (+/- 184,006)
test bench_pseurand_big_farm_direct   ... bench:     344,714 ns/iter (+/- 18,739)
test bench_pseurand_big_fnv           ... bench:   3,114,750 ns/iter (+/- 289,124)
test bench_pseurand_big_sip           ... bench:   1,462,424 ns/iter (+/- 87,814)
```