@bluss 2015-08-26
Commit 2 :smile:

```
test bench_pseurand_big_farm          ... bench:   2,576,059 ns/iter (+/- 61,059)
test bench_pseurand_big_farm_direct   ... bench:     324,533 ns/iter (+/- 24,064)
test bench_pseurand_big_fnv           ... bench:   3,120,601 ns/iter (+/- 129,750)
test bench_pseurand_big_sip           ... bench:   1,457,251 ns/iter (+/- 122,839)
```

Hehe it's a pretty tedious thing.

So.. an amazing number of boundary checks are already optimized out. We can try to chase the last ones, and it has an actual difference, for example functions use less registers if they have no panic branches. So it definitely matters to remove these in tight loops.