@bluss 2015-08-26
There's one last one that I don't know how to solve by "tricks" in safe code alone.

This [exact one on line 37](https://github.com/seiflotfy/rust-farmhash/blob/master/src/farmhashna_shared.rs#L37), it's the only bounds check in that function that's not optimized out in my profile. It's weird how it works out (and is it fragile, may change with compiler version?).