@bluss 2015-08-26
Ok, commit 3 addresses that one. Maybe that commit is silly, don't know what you think. Improvement on `bench_lorem_farm_direct` is only 3-4%.